api alpha
Moved folders around into andrew's organization

add Communication Layer in Napi
This commit includes the ability for javascript to both send and recieve
packets from C++. This is currently done by sending an agreed upon
packet structure between the two code bases. The C++ side is not yet
linked to the FPGA/Triggering pipeline, but can be configured to return
points (ie a sine-wave) for graph testing purposes. To do this ask
Daniel for help.

Worker Pool Created
Added in the thread worker pool that will take request packets from JavaScript,
Process the request, then send a response back to javascript.

Changed the packet format to make it easier to use. Documented the changes and it's structure in the code

Quick Fix

Fixed a bug that seemly caused js to crash for no apperantly reason. Added the FTDI lib to github so i dont have to hunt it down every time