Fixed multiple trigger next to eachother issue
If 2 triggers were within the same 64 bit section, the log2 would grab
the one occuring later in time, rather than earlier. Fixed by reversing
the order of the trigger bits and subtracting the log2 pos from 63.

Changed cli to allow passing a csv file to --t1. Had to remove the
option of running multiple tests at once because it was implemented in a
simple for loop and would read read the csv as a new argument instead of
part of the previous one.