Commit to add Communication Layer in Napi
This commit includes the ability for javascript to both send and recieve
packets from C++. This is currently done by sending an agreed upon
packet structure between the two code bases. The C++ side is not yet
linked to the FPGA/Triggering pipeline, but can be configured to return
points (ie a sine-wave) for graph testing purposes. To do this ask
Daniel for help.